Rename Http to Client and HttptMultiWorkspace to ClientMultiWorkspace
