+++ spec/request_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Seam::Http do
+RSpec.describe Seam::Client do
   let(:client) { Seam.new(api_key: "seam_some_api_key") }
 
   describe "Exceptions" do
