+++ spec/seam_client/client_multi_workspace_spec.rb
@@ -2,14 +2,14 @@
 
 require "spec_helper"
 
-RSpec.describe Seam::HttpMultiWorkspace do
+RSpec.describe Seam::ClientMultiWorkspace do
   let(:personal_access_token) { "seam_at_12345" }
   let(:endpoint) { "https://example.com/api" }
   let(:client) { described_class.from_personal_access_token(personal_access_token, endpoint: endpoint) }
 
   describe ".from_personal_access_token" do
     it "creates a new instance with the given token and endpoint" do
-      expect(client).to be_a(Seam::HttpMultiWorkspace)
+      expect(client).to be_a(Seam::ClientMultiWorkspace)
       expect(client.instance_variable_get(:@auth_headers)).to include("authorization" => "Bearer #{personal_access_token}")
       expect(client.instance_variable_get(:@endpoint)).to eq(endpoint)
     end
@@ -39,9 +39,17 @@ RSpec.describe Seam::HttpMultiWorkspace do
 
   describe "token format validation" do
     it "raises SeamInvalidTokenError for invalid token formats" do
-      expect { described_class.from_personal_access_token("invalid_token") }.to raise_error(Http::SeamInvalidTokenError, /Unknown/)
-      expect { described_class.from_personal_access_token("seam_apikey_token") }.to raise_error(Http::SeamInvalidTokenError, /Unknown/)
-      expect { described_class.from_personal_access_token("seam_cst") }.to raise_error(Http::SeamInvalidTokenError, /Client Session Token/)
+      expect do
+        described_class.from_personal_access_token("invalid_token")
+      end.to raise_error(Http::SeamInvalidTokenError, /Unknown/)
+      expect do
+        described_class.from_personal_access_token("seam_apikey_token")
+      end.to raise_error(Http::SeamInvalidTokenError,
+                         /Unknown/)
+      expect do
+        described_class.from_personal_access_token("seam_cst")
+      end.to raise_error(Http::SeamInvalidTokenError,
+                         /Client Session Token/)
       expect { described_class.from_personal_access_token("ey") }.to raise_error(Http::SeamInvalidTokenError, /JWT/)
     end
   end
